// src/types/WhitelistModuleHelper.d.ts
interface WhitelistModuleHelperInterface extends ethers.utils.Interface {
  functions: {
    "executeAction(uint256,address[],uint256[],bytes[])": FunctionFragment;
    "whitelist(address)": FunctionFragment;
  };

  encodeFunctionData(
    functionFragment: "executeAction",
    values: [BigNumberish, string[], BigNumberish[], BytesLike[]]
  ): string;
  encodeFunctionData(functionFragment: "whitelist", values: [string]): string;

  decodeFunctionResult(
    functionFragment: "executeAction",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "whitelist", data: BytesLike): Result;

  events: {};
}

export class WhitelistModuleHelper extends Contract {
  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  on(event: EventFilter | string, listener: Listener): this;
  once(event: EventFilter | string, listener: Listener): this;
  addListener(eventName: EventFilter | string, listener: Listener): this;
  removeAllListeners(eventName: EventFilter | string): this;
  removeListener(eventName: any, listener: Listener): this;

  interface: WhitelistModuleHelperInterface;

  functions: {
    executeAction(
      _proposalId: BigNumberish,
      _actionTos: string[],
      _actionValues: BigNumberish[],
      _actionDatas: BytesLike[],
      overrides?: Overrides
    ): Promise<ContractTransaction>;

    "executeAction(uint256,address[],uint256[],bytes[])"(
      _proposalId: BigNumberish,
      _actionTos: string[],
      _actionValues: BigNumberish[],
      _actionDatas: BytesLike[],
      overrides?: Overrides
    ): Promise<ContractTransaction>;

    whitelist(
      arg0: string,
      overrides?: CallOverrides
    ): Promise<{
      0: boolean;
    }>;

    "whitelist(address)"(
      arg0: string,
      overrides?: CallOverrides
    ): Promise<{
      0: boolean;
    }>;
  };

  executeAction(
    _proposalId: BigNumberish,
    _actionTos: string[],
    _actionValues: BigNumberish[],
    _actionDatas: BytesLike[],
    overrides?: Overrides
  ): Promise<ContractTransaction>;

  "executeAction(uint256,address[],uint256[],bytes[])"(
    _proposalId: BigNumberish,
    _actionTos: string[],
    _actionValues: BigNumberish[],
    _actionDatas: BytesLike[],
    overrides?: Overrides
  ): Promise<ContractTransaction>;

  whitelist(arg0: string, overrides?: CallOverrides): Promise<boolean>;

  "whitelist(address)"(
    arg0: string,
    overrides?: CallOverrides
  ): Promise<boolean>;

  callStatic: {
    executeAction(
      _proposalId: BigNumberish,
      _actionTos: string[],
      _actionValues: BigNumberish[],
      _actionDatas: BytesLike[],
      overrides?: CallOverrides
    ): Promise<void>;

    "executeAction(uint256,address[],uint256[],bytes[])"(
      _proposalId: BigNumberish,
      _actionTos: string[],
      _actionValues: BigNumberish[],
      _actionDatas: BytesLike[],
      overrides?: CallOverrides
    ): Promise<void>;

    whitelist(arg0: string, overrides?: CallOverrides): Promise<boolean>;

    "whitelist(address)"(
      arg0: string,
      overrides?: CallOverrides
    ): Promise<boolean>;
  };

  filters: {};

  estimateGas: {
    executeAction(
      _proposalId: BigNumberish,
      _actionTos: string[],
      _actionValues: BigNumberish[],
      _actionDatas: BytesLike[],
      overrides?: Overrides
    ): Promise<BigNumber>;

    "executeAction(uint256,address[],uint256[],bytes[])"(
      _proposalId: BigNumberish,
      _actionTos: string[],
      _actionValues: BigNumberish[],
      _actionDatas: BytesLike[],
      overrides?: Overrides
    ): Promise<BigNumber>;

    whitelist(arg0: string, overrides?: CallOverrides): Promise<BigNumber>;

    "whitelist(address)"(
      arg0: string,
      overrides?: CallOverrides
    ): Promise<BigNumber>;
  };

  populateTransaction: {
    executeAction(
      _proposalId: BigNumberish,
      _actionTos: string[],
      _actionValues: BigNumberish[],
      _actionDatas: BytesLike[],
      overrides?: Overrides
    ): Promise<PopulatedTransaction>;

    "executeAction(uint256,address[],uint256[],bytes[])"(
      _proposalId: BigNumberish,
      _actionTos: string[],
      _actionValues: BigNumberish[],
      _actionDatas: BytesLike[],
      overrides?: Overrides
    ): Promise<PopulatedTransaction>;

    whitelist(
      arg0: string,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;

    "whitelist(address)"(
      arg0: string,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;
  };
}